+++ src/op/iter_edges.rs
@@ -7,31 +7,11 @@
 //! use graaf::op::IterEdges;
 //!
 //! let graph = [vec![1, 2], vec![0, 2, 3], vec![0, 1, 3], vec![1, 2]];
-//! let mut iter = graph.iter_edges(0);
 //!
-//! assert_eq!(iter.next(), Some(1));
-//! assert_eq!(iter.next(), Some(2));
-//! assert_eq!(iter.next(), None);
-//!
-//! let mut iter = graph.iter_edges(1);
-//!
-//! assert_eq!(iter.next(), Some(0));
-//! assert_eq!(iter.next(), Some(2));
-//! assert_eq!(iter.next(), Some(3));
-//! assert_eq!(iter.next(), None);
-//!
-//! let mut iter = graph.iter_edges(2);
-//!
-//! assert_eq!(iter.next(), Some(0));
-//! assert_eq!(iter.next(), Some(1));
-//! assert_eq!(iter.next(), Some(3));
-//! assert_eq!(iter.next(), None);
-//!
-//! let mut iter = graph.iter_edges(3);
-//!
-//! assert_eq!(iter.next(), Some(1));
-//! assert_eq!(iter.next(), Some(2));
-//! assert_eq!(iter.next(), None);
+//! assert!(graph.iter_edges(0).eq([1, 2].into_iter()));
+//! assert!(graph.iter_edges(1).eq([0, 2, 3].into_iter()));
+//! assert!(graph.iter_edges(2).eq([0, 1, 3].into_iter()));
+//! assert!(graph.iter_edges(3).eq([1, 2].into_iter()));
 //! ```
 //!
 //! The order of the edges is not guaranteed for, e.g., `Vec<HashSet<_>>`:
@@ -98,31 +78,11 @@ use {
 /// use graaf::op::IterEdges;
 ///
 /// let graph = [vec![1, 2], vec![0, 2, 3], vec![0, 1, 3], vec![1, 2]];
-/// let mut iter = graph.iter_edges(0);
 ///
-/// assert_eq!(iter.next(), Some(1));
-/// assert_eq!(iter.next(), Some(2));
-/// assert_eq!(iter.next(), None);
-///
-/// let mut iter = graph.iter_edges(1);
-///
-/// assert_eq!(iter.next(), Some(0));
-/// assert_eq!(iter.next(), Some(2));
-/// assert_eq!(iter.next(), Some(3));
-/// assert_eq!(iter.next(), None);
-///
-/// let mut iter = graph.iter_edges(2);
-///
-/// assert_eq!(iter.next(), Some(0));
-/// assert_eq!(iter.next(), Some(1));
-/// assert_eq!(iter.next(), Some(3));
-/// assert_eq!(iter.next(), None);
-///
-/// let mut iter = graph.iter_edges(3);
-///
-/// assert_eq!(iter.next(), Some(1));
-/// assert_eq!(iter.next(), Some(2));
-/// assert_eq!(iter.next(), None);
+/// assert!(graph.iter_edges(0).eq([1, 2].into_iter()));
+/// assert!(graph.iter_edges(1).eq([0, 2, 3].into_iter()));
+/// assert!(graph.iter_edges(2).eq([0, 1, 3].into_iter()));
+/// assert!(graph.iter_edges(3).eq([1, 2].into_iter()));
 /// ```
 ///
 /// The order of the edges is not guaranteed for, e.g., `Vec<HashSet<_>>`:
@@ -293,31 +253,10 @@ mod tests {
 
     macro_rules! test_iter_edges_stable {
         ($graph:expr) => {
-            let mut iter = $graph.iter_edges(0);
-
-            assert_eq!(iter.next(), Some(1));
-            assert_eq!(iter.next(), Some(2));
-            assert_eq!(iter.next(), None);
-
-            let mut iter = $graph.iter_edges(1);
-
-            assert_eq!(iter.next(), Some(0));
-            assert_eq!(iter.next(), Some(2));
-            assert_eq!(iter.next(), Some(3));
-            assert_eq!(iter.next(), None);
-
-            let mut iter = $graph.iter_edges(2);
-
-            assert_eq!(iter.next(), Some(0));
-            assert_eq!(iter.next(), Some(1));
-            assert_eq!(iter.next(), Some(3));
-            assert_eq!(iter.next(), None);
-
-            let mut iter = $graph.iter_edges(3);
-
-            assert_eq!(iter.next(), Some(1));
-            assert_eq!(iter.next(), Some(2));
-            assert_eq!(iter.next(), None);
+            assert!($graph.iter_edges(0).eq([1, 2].into_iter()));
+            assert!($graph.iter_edges(1).eq([0, 2, 3].into_iter()));
+            assert!($graph.iter_edges(2).eq([0, 1, 3].into_iter()));
+            assert!($graph.iter_edges(3).eq([1, 2].into_iter()));
         };
     }